# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Boncos — Nota Kalcer API",
    "version": "1.0.0",
    "description": "Dokumentasi REST API interaktif untuk **Boncos**, aplikasi pencatatan keuangan pribadi neobrutalisme Gen Z berbasis Next.js 15 dan Firebase.\n\n### 🛡️ Catatan Keamanan & Batasan:\n- **Batas Waktu Eksekusi (Timeout):** Setiap endpoint memiliki batas waktu maksimal **30 detik** (`maxDuration = 30s`), mengembalikan status HTTP 504 jika proses melebihi batas waktu.\n- **Isolasi User:** Setiap pengguna diisolasi via `uid` dari session cookie Firebase (`__session`) atau kode unik ingest (`key`). Tidak ada parameter `uid` dari request body atau query.\n- **Uang Selalu Integer:** Nominal uang selalu bilangan bulat rupiah tanpa desimal.\n- **Enkripsi Telegram:** Bot Token Telegram dienkripsi dengan AES-256-GCM sebelum disimpan di database dan tidak pernah diekspos ke klien."
  },
  "servers": [
    {
      "url": "/",
      "description": "Server Aktif (Current Origin)"
    },
    {
      "url": "http://localhost:3000",
      "description": "Server Development Lokal"
    }
  ],
...
  "paths": {
...
    "/api/telegram": {
      "get": {
        "tags": ["Telegram Settings"],
        "summary": "Ambil Konfigurasi Notifikasi Telegram",
        "description": "Mengambil konfigurasi notifikasi Telegram pengguna. Token bot asli tidak pernah dikembalikan ke klien; hanya petunjuk 4 karakter terakhir (`bot_token_hint`).",
        "security": [{ "SessionCookie": [] }],
        "responses": {
          "200": {
            "description": "Pengaturan Telegram ditemukan",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": { "type": "boolean", "example": true },
                    "data": { "$ref": "#/components/schemas/TelegramSettingsView" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Belum login",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Telegram Settings"],
        "summary": "Simpan Konfigurasi Telegram",
        "description": "Menyimpan pengaturan notifikasi Telegram. Memvalidasi token ke Telegram API (`getMe`), mengenkripsi token menggunakan kunci rahasia server (AES-256-GCM), lalu menyimpannya ke Firestore.",
        "security": [{ "SessionCookie": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["enabled", "chat_id"],
                "properties": {
                  "enabled": { "type": "boolean", "example": true },
                  "chat_id": { "type": "string", "example": "-1001234567890" },
                  "bot_token": { "type": "string", "example": "123456789:ABCdefGHIjklMNOpqrSTUvwxYZ" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Pengaturan Telegram berhasil disimpan",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessResponse" }
              }
            }
          },
          "400": {
            "description": "Token bot Telegram atau Chat ID tidak valid",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Telegram Settings"],
        "summary": "Hapus Konfigurasi Telegram",
        "description": "Menghapus konfigurasi dan token Telegram pengguna dari Firestore.",
        "security": [{ "SessionCookie": [] }],
        "responses": {
          "200": {
            "description": "Pengaturan Telegram berhasil dihapus",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessResponse" }
              }
            }
          }
        }
      }
    },
...
    "/api/scan": {
      "post": {
        "tags": ["Scan & Webhook"],
        "summary": "Scan Struk & Auto Catat via Kode Unik (Headless Ingestion)",
        "description": "Endpoint publik tanpa autentikasi cookie session. Menerima gambar struk dan kode unik rahasia user (`key`). Gambar diproses OCR di memori server secara aman (tanpa disimpan ke storage/disk). Jika berhasil, pengeluaran otomatis dicatat dan dilaporkan ke Telegram (jika terhubung). Jika gagal, mengembalikan rincian error dan mengirim notifikasi kegagalan ke Telegram.",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["key", "image"],
                "properties": {
                  "key": {
                    "type": "string",
                    "description": "Kode unik pengguna yang diperoleh dari menu Pengaturan -> Kode Unik & Scan API",
                    "example": "bnc_9a7d8e6c4b2a1f0e3d5c7b9a"
                  },
                  "image": {
                    "type": "string",
                    "format": "binary",
                    "description": "Berkas gambar struk belanja (JPEG, PNG, WebP maksimal 10MB)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Struk berhasil dipindai dan dicatat ke sistem",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": { "type": "boolean", "example": true },
                    "message": { "type": "string", "example": "Struk berhasil discan dan dicatat ke pengeluaran." },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "string", "example": "exp_abc123" },
                        "name": { "type": "string", "example": "Kopi Nako" },
                        "amount": { "type": "integer", "example": 102300 },
                        "expense_date": { "type": "string", "example": "2026-09-13" },
                        "category_id": { "type": "string", "example": "food" },
                        "confidence": { "type": "number", "example": 0.89 }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "File gambar tidak disertakan atau ukuran melebihi batas",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Kode unik (key) tidak valid atau tidak disertakan",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "422": {
            "description": "Struk tidak terbaca atau nominal tidak ditemukan",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "500": {
            "description": "Terjadi kesalahan sistem saat pemrosesan",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "SessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__session",
        "description": "Firebase Auth Session Cookie terverifikasi (HttpOnly)"
      }
    },
    "schemas": {
...
      "TelegramSettingsView": {
        "type": "object",
        "properties": {
          "enabled": { "type": "boolean", "example": true },
          "chat_id": { "type": "string", "example": "-1001234567890" },
          "bot_token_hint": { "type": "string", "example": "...a1b2" }
        }
      },
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "ok": { "type": "boolean", "example": true }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "ok": { "type": "boolean", "example": false },
          "error": { "type": "string", "example": "INVALID_INPUT" },
          "message": { "type": "string", "example": "Pesan error berbahasa Indonesia yang ramah" }
        }
      }
    }
  }
}